Rename enteties to entities
