+++ member-service/src/main/java/com/hedvig/memberservice/entities/CollectResponse.java
@@ -1,4 +1,4 @@
-package com.hedvig.memberservice.enteties;
+package com.hedvig.memberservice.entities;
 
 import com.hedvig.external.bankID.bankIdRestTypes.CollectStatus;
 import javax.persistence.Embeddable;
