+++ member-service/src/main/java/com/hedvig/memberservice/entities/SignSessionRepository.java
@@ -1,4 +1,4 @@
-package com.hedvig.memberservice.enteties;
+package com.hedvig.memberservice.entities;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.Query;
