+++ member-service/src/main/java/com/hedvig/memberservice/entities/SignStatus.java
@@ -1,4 +1,4 @@
-package com.hedvig.memberservice.enteties;
+package com.hedvig.memberservice.entities;
 
 public enum SignStatus {
