+++ member-service/src/main/java/com/hedvig/memberservice/services/SigningService.java
@@ -7,10 +7,10 @@ import static org.quartz.TriggerBuilder.newTrigger;
 import com.hedvig.external.bankID.bankIdRestTypes.BankIdRestError;
 import com.hedvig.external.bankID.bankIdRestTypes.CollectStatus;
 import com.hedvig.external.bankID.bankIdRestTypes.OrderResponse;
-import com.hedvig.memberservice.enteties.CollectResponse;
-import com.hedvig.memberservice.enteties.SignSession;
-import com.hedvig.memberservice.enteties.SignSessionRepository;
-import com.hedvig.memberservice.enteties.SignStatus;
+import com.hedvig.memberservice.entities.CollectResponse;
+import com.hedvig.memberservice.entities.SignSession;
+import com.hedvig.memberservice.entities.SignSessionRepository;
+import com.hedvig.memberservice.entities.SignStatus;
 import com.hedvig.memberservice.externalApi.botService.BotService;
 import com.hedvig.memberservice.externalApi.botService.dto.UpdateUserContextDTO;
 import com.hedvig.memberservice.externalApi.productsPricing.ProductApi;
