+++ member-service/src/main/java/com/hedvig/memberservice/services/member/dto/MemberSignResponse.java
@@ -1,7 +1,7 @@
 package com.hedvig.memberservice.services.member.dto;
 
 import com.hedvig.external.bankID.bankIdRestTypes.OrderResponse;
-import com.hedvig.memberservice.enteties.SignStatus;
+import com.hedvig.memberservice.entities.SignStatus;
 import lombok.Value;
 import org.springframework.lang.NonNull;
