+++ member-service/src/main/java/com/hedvig/memberservice/web/v2/dto/WebSignResponse.java
@@ -1,7 +1,7 @@
 package com.hedvig.memberservice.web.v2.dto;
 
 import com.hedvig.external.bankID.bankIdRestTypes.OrderResponse;
-import com.hedvig.memberservice.enteties.SignStatus;
+import com.hedvig.memberservice.entities.SignStatus;
 import lombok.NonNull;
 import lombok.Value;
