+++ member-service/src/test/java/com/hedvig/memberservice/services/SigningServiceTest.java
@@ -18,9 +18,9 @@ import com.hedvig.external.bankID.bankIdRestTypes.CollectResponse;
 import com.hedvig.external.bankID.bankIdRestTypes.CollectStatus;
 import com.hedvig.external.bankID.bankIdRestTypes.CompletionData;
 import com.hedvig.external.bankID.bankIdRestTypes.OrderResponse;
-import com.hedvig.memberservice.enteties.SignSession;
-import com.hedvig.memberservice.enteties.SignSessionRepository;
-import com.hedvig.memberservice.enteties.SignStatus;
+import com.hedvig.memberservice.entities.SignSession;
+import com.hedvig.memberservice.entities.SignSessionRepository;
+import com.hedvig.memberservice.entities.SignStatus;
 import com.hedvig.memberservice.externalApi.botService.BotService;
 import com.hedvig.memberservice.externalApi.productsPricing.ProductApi;
 import com.hedvig.memberservice.externalApi.productsPricing.dto.ProductToSignStatusDTO;
